Enable button create_new/submit Correct in views and bill_new
The value of the "supply" field was not passed to the form.
This field is "No" by default in model.
But if it is not on the page of creating a new order -
this field is not in the passed parameters.

I had to create this field on the page.
And then the button "submit" worked.